commands: move unsquash code to unsquash.rs
